#190 Prepare for versioned release
