+++ szconfigmanager/szconfigmanager_examples_test.go
@@ -44,6 +44,66 @@ func ExampleSzconfigmanager_CreateConfigFromConfigID() {
 	// Output: {"G2_CONFIG":{"CFG_ATTR":[{"ATTR_CLASS":"ADDRESS","ATTR_CODE":"ADDR_CITY","ATTR_ID":1608,...
 }
 
+func ExampleSzconfigmanager_CreateConfigFromString() {
+	// For more information, visit
+	// https://github.com/senzing-garage/sz-sdk-go-mock/blob/main/szconfigmanager/szconfigmanager_examples_test.go
+	ctx := context.TODO()
+	szAbstractFactory := getSzAbstractFactory(ctx)
+
+	szConfigManager, err := szAbstractFactory.CreateConfigManager(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	szConfigFromTemplate, err := szConfigManager.CreateConfigFromTemplate(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	configDefinitionFromTemplate, err := szConfigFromTemplate.Export(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	szConfig, err := szConfigManager.CreateConfigFromString(ctx, configDefinitionFromTemplate)
+	if err != nil {
+		handleError(err)
+	}
+
+	configDefinition, err := szConfig.Export(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	fmt.Println(jsonutil.Truncate(configDefinition, 7))
+	// Output: {"G2_CONFIG":{"CFG_ATTR":[{"ATTR_CLASS":"ADDRESS","ATTR_CODE":"ADDR_CITY","ATTR_ID":1608,...
+}
+
+func ExampleSzconfigmanager_CreateConfigFromTemplate() {
+	// For more information, visit
+	// https://github.com/senzing-garage/sz-sdk-go-mock/blob/main/szconfigmanager/szconfigmanager_examples_test.go
+	ctx := context.TODO()
+	szAbstractFactory := getSzAbstractFactory(ctx)
+
+	szConfigManager, err := szAbstractFactory.CreateConfigManager(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	szConfig, err := szConfigManager.CreateConfigFromTemplate(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	configDefinition, err := szConfig.Export(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	fmt.Println(jsonutil.Truncate(configDefinition, 7))
+	// Output: {"G2_CONFIG":{"CFG_ATTR":[{"ATTR_CLASS":"ADDRESS","ATTR_CODE":"ADDR_CITY","ATTR_ID":1608,...
+}
+
 func ExampleSzconfigmanager_GetConfigs() {
 	// For more information, visit
 	// https://github.com/senzing-garage/sz-sdk-go-mock/blob/main/szconfigmanager/szconfigmanager_examples_test.go
@@ -164,6 +224,50 @@ func ExampleSzconfigmanager_ReplaceDefaultConfigID() {
 	// Output:
 }
 
+func ExampleSzconfigmanager_SetDefaultConfig() {
+	// For more information, visit
+	// https://github.com/senzing-garage/sz-sdk-go-mock/blob/main/szconfigmanager/szconfigmanager_examples_test.go
+	ctx := context.TODO()
+	szAbstractFactory := getSzAbstractFactory(ctx)
+
+	szConfigManager, err := szAbstractFactory.CreateConfigManager(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	defaultConfigID, err := szConfigManager.GetDefaultConfigID(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	szConfig, err := szConfigManager.CreateConfigFromConfigID(ctx, defaultConfigID)
+	if err != nil {
+		handleError(err)
+	}
+
+	dataSourceCode := "GO_TEST"
+
+	_, err = szConfig.AddDataSource(ctx, dataSourceCode)
+	if err != nil {
+		handleError(err)
+	}
+
+	configComment := "Added datasource: " + dataSourceCode
+
+	configDefinition, err := szConfig.Export(ctx)
+	if err != nil {
+		handleError(err)
+	}
+
+	configID, err := szConfigManager.SetDefaultConfig(ctx, configDefinition, configComment)
+	if err != nil {
+		handleError(err)
+	}
+
+	fmt.Println(configID > 0) // Dummy output.
+	// Output: false
+}
+
 func ExampleSzconfigmanager_SetDefaultConfigID() {
 	// For more information, visit
 	// https://github.com/senzing-garage/sz-sdk-go-mock/blob/main/szconfigmanager/szconfigmanager_examples_test.go
@@ -227,14 +331,4 @@ func ExampleSzconfigmanager_GetObserverOrigin() {
 
 	fmt.Println(result)
 	// Output: Machine: nn; Task: UnitTest
-}
-
-// ----------------------------------------------------------------------------
-// Helper functions
-// ----------------------------------------------------------------------------
-
-func handleError(err error) {
-	if err != nil {
-		fmt.Println("Error:", err)
-	}
 }
